// ledger-engine-frontend/src/lib/api/types/api.d.ts
export interface components {
    schemas: {
// ...
        NotificationResponse: {
            /** Format: uuid */
            notification_id?: string;
            /** Format: uuid */
            user_id?: string;
            /** @enum {string} */
            type?: "TOPUP_COMPLETED" | "P2P_RECEIVED" | "BILL_PAID" | "SECURITY_ALERT";
            title?: string;
            message?: string;
            is_read?: boolean;
            /** Format: date-time */
            created_at?: string;
        };
        WebKycSubmitRequest: {
            /** Format: uuid */
            userId: string;
            documentType: string;
            documentNumber: string;
        };
        WebRegisterRequest: {
            /** Format: email */
            email: string;
            phoneNumber: string;
            firstName: string;
            lastName: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}